File handles changed to int, since the tape is -1

+++ src/lib/libbasic/CLOSE.c
@@ -1,7 +1,7 @@
 
 #include <basic.h>
 
-int CLOSE(byte f)
+int CLOSE(int f)
 {
     return systemf("CLOSE #%d", f);
 }
